driver/plx905x.c: Fix EEPROM access for PCI9056
For the PLX PCI9056, bit 31 of the CNTRL register needs to be
manipulated in the same way as for the PCI9656 to avoid contention on
the multiplexed EEDI/EEDO bus.   Change the initialization of
`plx905x_device.cntrl_eemask` for PCI9056 to be the same as is used for
PCI9656.

Also, rename the macro PLX9656_EEMASK to PLX9056_EEMASK.  It expands to
the `plx905x_device.cntrl_eemask` value for PCI9056 and PCI9656.