driver/kcompat.h: Use MEV's latest version - defines 'fallthrough' macro
Linux kernel version 5.14 expects code that falls through to a 'case' or
'default' label in a switch statement to use the 'fallthrough;'
statement to avoid a warning.  The special comment contents no longer
work.  The 'fallthrough' macro was first defined in kernel version 5.4.
Add a dummy macro if it is missing.

Code should use 'fallthrough;' and a special comment on the following
line to avoid '-Wimplicit-fallthrough' warnings on all kernel versions.