driver/kcompat.h: Fix for RHEL back-port of class_create(name) to kernel 5.14
Change pre-processor tests for redefining `class_create(name)` for
kernel versions prior to 6.4 due to Red Hat kernel API back-port of the
6.4 version of `class_create(name)` to later 5.14 RHEL kernels.  If we
declare `kcompat_class_create(owner, name)` as a function, then also
define it as a macro of the same name.  For kernel versions prior to 6.4
but at least 2.6.27, only define `kcompat_class_create(owner, name)` as
a macro if the `class_create` macro is defined.  Only redefine
`class_create(name)` as a macro if the kernel version is less than 6.4
and we have defined `kcompat_class_create` as a macro.