driver/kcompat.h: Change __list_splice() compatibility
In kernel versions prior to 2.6.27, __list_splice() had two parameters,
being a pointer to the control node of the list to be moved (which is
assumed to be non-empty) and a pointer to the control node of the target
list.  In later kernel versions, there are three parameters and the
pointer to the control node of the target list has been replaced with
pointers to the nodes either side of where the list is to be inserted.

For kernel versions prior to 2.6.27, define __kcompat_list_splice() with
three parameters and define __list_splice() as a macro with three
parameters that calls __kcompat_list_splice().  This is used by our
compatible versions of list_splice(), list_splice_init(),
list_splice_tail(), and list_splice_tail_init().