driver: Define kcompat_access_ok() function
Define kcompat_access_ok() as a static inline function with two
parameters to be used in place of the two parameter version of
access_ok() for kernel versions 5.0 and later.  For earlier kernel
versions, it will call the three parameter version of access_ok() with
the first parameter always set to VERIFY_WRITE.  This will fail for some
old 386 processors if the memory is not in fact writable.

Replace calls to access_ok() with calls to kcompat_access_ok() in
plx905x.c.